requirements.txt
tmdbsimple
python-dotenv
sendgrid==6.6.0
